+++ ricaun.RevitTest.Tests/TestsRevitTask.cs
@@ -2,6 +2,7 @@
 using NUnit.Framework;
 using ricaun.Revit.UI.Tasks;
 using System;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
 
@@ -52,6 +53,20 @@ namespace ricaun.RevitTest.Tests
         private bool InContext(UIApplication uiapp)
         {
             return !(uiapp.ActiveAddInId is null);
+        }
+
+        [Test]
+        public void GetAssemblies()
+        {
+            var assemblies = AppDomain.CurrentDomain.GetAssemblies().Where(assembly => typeof(TestsRevitTask).Assembly.FullName == assembly.FullName);
+            foreach (var assembly in assemblies)
+            {
+                Console.WriteLine($"{assembly.GetName().Name} | {(!string.IsNullOrEmpty(assembly.Location) ? System.IO.Path.GetFileName(assembly.Location) : string.Empty)}");
+            }
+#if !DEBUG
+            // Check if the assembly is only loaded once
+            Assert.AreEqual(1, assemblies.Count());
+#endif
         }
 
         [Test]