+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Application
 - Update to use `AssemblyMetadata` to set configuration.
 - Update `TestExecuteUtils` to use `ConfigurationMetadata`.
+- Update `Log` file name.
+- Update `ricaun.NUnit` to `1.4.1`.
 
 ## [1.7.0] / 2024-11-02 - 2024-12-03
 ### Features